Adding EmitC support for repeat op.

+++ test/ttmlir/Dialect/TTIR/data_movement/broadcast/ttir_broadcast_negative_rank.mlir
@@ -6,7 +6,7 @@ module {
   func.func @main(%arg0: tensor<1x16x32xf32>, %arg1: tensor<1x1x32xf32>) -> tensor<1x16x32xf32> {
     // CHECK: error: 'ttir.broadcast' op Input tensor rank should match output tensor rank.
     %0 = tensor.empty() : tensor<1x16x32xf32>
-    %1 = "ttir.broadcast"(%arg1, %0) <{broadcast_dimensions = array<i32 : 1, 16>}> : (tensor<1x1x32xf32>, tensor<1x16x32xf32>) -> tensor<1x16x32xf32>
+    %1 = "ttir.broadcast"(%arg1, %0) <{broadcast_dimensions = array<i64 : 1, 16>}> : (tensor<1x1x32xf32>, tensor<1x16x32xf32>) -> tensor<1x16x32xf32>
     %2 = tensor.empty() : tensor<1x16x32xf32>
     %3 = "ttir.multiply"(%arg0, %1, %2) <{operandSegmentSizes = array<i32: 2, 1>}> : (tensor<1x16x32xf32>, tensor<1x16x32xf32>, tensor<1x16x32xf32>) -> tensor<1x16x32xf32>
     return %3 : tensor<1x16x32xf32>
